Refactor Event and Sample form DataItem.cs
Pulled out these classes into their own files for clarity

Added copyright claim to top of other DataItem files
